Put db upgraders for 2-to-3 and 3-to-4 in their own modules
This makes the __init__ of sqlite.db_upgrades more concise, and also
serves as an example of how to encapsulate code needed by a
particular upgrader function.
